Rearrange vehicle order to correct part load errors
Vehicle parts must be ordered with y-values descending from -1 and
ascending from 0. Otherwise, parts are unable to be placed due to
not having a frame to attach to. The previous vehicle reformatter
did not handle this correctly, and I unknowingly introduced a large
number of errors into the codebase when I ran the script. This
commit removes those errors.