Update to vanilla standards + mod additions
This updates the requirements to vanilla DDA standards + mod additions as seems to have been the original intention. The only thing I'm uncertain of is the "survivor_belt" in wire cutting...I think that's handled differently now, but I'm not certain. And I left the toolset in wire cutting because it makes sense. All else is vanilla + the original mod.