Update base requirements to vanilla + additions
addition, really. These seem to overwrite so this mod directly conflicts with Advanced Gear Technologies at least as far as the glare number goes. I didn't test to see if the list of items gets extended or just overwritten. It's hard to tell which way that will flop and I'm going to save myself a headache and just make another compatibility mod to combine these things for my sanity if nothing else...